fix subset label bug

+++ include/FedTree/FL/partition.h
@@ -33,8 +33,8 @@ public:
 
     void hybrid_partition_practical(const DataSet &dataset, const int n_parties,
                                     vector<SyncArray<bool>> &feature_map, vector<DataSet> &subsets,
-                                    int ins_gau_mean_factor, int ins_gau_sigma,
-                                    int fea_gau_mean_factor, int fea_gau_sigma, int seed = 42);
+                                    int ins_gau_mean_factor, float ins_gau_sigma,
+                                    int fea_gau_mean_factor, float fea_gau_sigma, int seed = 42);
 };
 
 #endif //FEDTREE_PARTITION_H
